relocate resolve encoder utils
